updated h1-counter
went in and added in additional logic for handling ingress data from the
remote host. the assignment specifies that the data should be checked
using the correct chunk size. not sure if that means that the buffer
needs to be completely filled or if a best attempt needs to be made to
just use what was filled in a single recv call. either way I made it an
option to try and fill the buffer before checking it for html tags. if
that needs to be the default behavior then I can flip the default flag
value and change the getopt_long argument name for it.

also fixed a few other issues when attempting to bind to a specific
local address where the port was not being honored so I removed it and
just have it deal with the local ip if one is provided. if the local ip
is not provided then it will not attempt to do the bind and just create
the socket and connect.

modified the stdout prints to be behind a verbose flag that is now
specified as a command line argument.