testing updates
when we attempted to test the server against the script specified in the
writeup the peer clients were erroring out. turns out the server needs
to handle taking a port argument. the server can now be tested but
failing tests so this will mark a point at which the server works but is
failing tests.